requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
anyio==4.10.0
astunparse==1.6.3
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
dotenv==0.9.9
fastapi==0.116.1
flatbuffers==25.2.10
gast==0.6.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
h5py==3.14.0
httplib2==0.30.0
idna==3.10
joblib==1.5.2
keras==3.11.3
libclang==18.1.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.3
namex==0.1.0
numpy==2.3.2
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pillow==11.3.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
python-dotenv==1.1.1
requests==2.32.5
rich==14.1.0
rsa==4.9.1
scikit-learn==1.7.1
scipy==1.16.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
starlette==0.47.3
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.3
